refactor: use shared review component for user operator
